AMP-3801
git-svn-id: file:///svn/trunk@9052 9e8d75aa-d0cb-44a9-8341-a4b8ba1a38b4

+++ amp/WEB-INF/src/org/dgfoundation/amp/ar/cell/TotalCommitmentsAmountCell.java
@@ -3,6 +3,7 @@
  */
 package org.dgfoundation.amp.ar.cell;
 
+import java.math.BigDecimal;
 import java.util.Iterator;
 
 import org.dgfoundation.amp.ar.ArConstants;
@@ -37,7 +38,7 @@ public class TotalCommitmentsAmountCell extends AmountCell {
          * 
          * @return Returns the amount.
          */
-    public double getAmount() {
+    public BigDecimal getAmount() {
 		if (id != null)
 			return convert();
 		return Logic.getInstance().getCommitmentCalculator().calculateAmount(mergedCells);
