+++ amp/WEB-INF/src/org/digijava/module/aim/dbentity/AmpPhysicalComponentReport.java
@@ -1,5 +1,6 @@
 package org.digijava.module.aim.dbentity;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 import org.digijava.module.aim.util.FeaturesUtil;
@@ -17,9 +18,9 @@ public class AmpPhysicalComponentReport {
 			private Date signatureDate;
 			private Date actualStartDate;
 			private Date plannedCompletionDate;
-			private Double actualCommitment;
+			private BigDecimal actualCommitment;
 			private String commCurrencyCode;
-			private Double actualExpenditure;
+			private BigDecimal actualExpenditure;
 			private String expCurrencyCode;
 			private String status;	
 			private Date reportingDate;
@@ -71,8 +72,8 @@ public class AmpPhysicalComponentReport {
 			return signatureDate;
 		}
 
-		public Double getActualCommitment() {
-					 return FeaturesUtil.applyThousandsForVisibility(actualCommitment);
+		public BigDecimal getActualCommitment() {
+					 return FeaturesUtil.applyThousandsForVisibility(actualCommitment)	;
 		  }
 
 		   public String getCommCurrencyCode() {
@@ -80,7 +81,7 @@ public class AmpPhysicalComponentReport {
 		  }
 
 		 
-		  public Double getActualExpenditure() {
+		  public BigDecimal getActualExpenditure() {
 					 return FeaturesUtil.applyThousandsForVisibility(actualExpenditure);
 		  }
 
@@ -172,7 +173,7 @@ public class AmpPhysicalComponentReport {
 			signatureDate = date;
 		}
 
-		  public void setActualCommitment(Double actualCommitment) {
+		  public void setActualCommitment(BigDecimal actualCommitment) {
 					 this.actualCommitment = FeaturesUtil.applyThousandsForEntry(actualCommitment);
 		  }
 
@@ -181,7 +182,7 @@ public class AmpPhysicalComponentReport {
 		  }
 
 	
-		  public void setActualExpenditure(Double actualExpenditure) {
+		  public void setActualExpenditure(BigDecimal actualExpenditure) {
 					 this.actualExpenditure = FeaturesUtil.applyThousandsForEntry(actualExpenditure);
 		  }
